% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_math.R
\name{m_math}
\alias{m_math}
\title{Raise and Subtract}
\usage{
m_math(a, b, c)
}
\arguments{
\item{a}{number}

\item{b}{number}

\item{c}{number}
}
\value{
number
}
\description{
Raise one number to another and subtract a third.
}
\examples{
m_math(4,2,2)
}
